requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
passlib==1.7.4
flask-wtf==1.0.0
email-validator==1.1.3
SQLAlchemy==1.4.23
psycopg2-binary==2.9.3
